+++ SimulCasterRenderer/src/crossplatform/Skin.h
@@ -19,7 +19,7 @@ public:
 	{
 		std::vector<mat4> boneMatrices(64);
 
-		for(int i = 0; i < bones.size(); i++)
+		for(size_t i = 0; i < bones.size(); i++)
 		{
 			boneMatrices[i] = rootTransform.GetTransformMatrix() * bones[i]->GetGlobalTransform().GetTransformMatrix() * inverseBindMatrices[i].GetTransformMatrix();
 		}
